Remove redundant DBLE()/REAL() type conversions in TESTING
In ddrvst.f, ddrvst2stg.f, and dlatm4.f, DBLE() was applied to
DOUBLE PRECISION array elements assigned to DOUBLE PRECISION
variables ― a no-op conversion. Similarly, in sdrvst.f,
sdrvst2stg.f, and slatm4.f, REAL() was applied to REAL array
elements assigned to REAL variables.

Remove these 168 redundant calls. Integer-to-real conversions
(e.g., DBLE(N), DBLE(KLEN-1), REAL(MAX(1,N))) are retained as
they are semantically necessary.

+++ TESTING/EIG/slatm4.f
@@ -343,10 +343,10 @@
 *        Scale by AMAGN
 *
          DO 230 JD = KBEG, KEND
-            A( JD, JD ) = AMAGN*REAL( A( JD, JD ) )
+            A( JD, JD ) = AMAGN*A( JD, JD )
   230    CONTINUE
          DO 240 JD = ISDB, ISDE
-            A( JD+1, JD ) = AMAGN*REAL( A( JD+1, JD ) )
+            A( JD+1, JD ) = AMAGN*A( JD+1, JD )
   240    CONTINUE
 *
 *        If ISIGN = 1 or 2, assign random signs to diagonal and
@@ -354,13 +354,13 @@
 *
          IF( ISIGN.GT.0 ) THEN
             DO 250 JD = KBEG, KEND
-               IF( REAL( A( JD, JD ) ).NE.ZERO ) THEN
+               IF( A( JD, JD ).NE.ZERO ) THEN
                   IF( SLARAN( ISEED ).GT.HALF )
      $               A( JD, JD ) = -A( JD, JD )
                END IF
   250       CONTINUE
             DO 260 JD = ISDB, ISDE
-               IF( REAL( A( JD+1, JD ) ).NE.ZERO ) THEN
+               IF( A( JD+1, JD ).NE.ZERO ) THEN
                   IF( SLARAN( ISEED ).GT.HALF )
      $               A( JD+1, JD ) = -A( JD+1, JD )
                END IF
